requirements.txt
discord.py==2.3.2
feedparser==6.0.10
python-dotenv==1.0.0 